+++ src/Frontend/Dfe.Complete/Pages/Projects/TaskList/Tasks/HandoverWithDeliveryOfficerTask/HandoverWithDeliveryOfficerTask.cshtml.cs
@@ -1,4 +1,9 @@
+using Dfe.Complete.Application.Projects.Commands.UpdateProject; 
+using Dfe.Complete.Constants; 
 using Dfe.Complete.Domain.Enums;
+using Dfe.Complete.Domain.ValueObjects;
+using Dfe.Complete.Extensions;
+using Dfe.Complete.Models;
 using MediatR;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
@@ -20,9 +25,37 @@ public class HandoverWithDeliveryOfficerTaskModel(ISender sender, IAuthorization
     [BindProperty(Name = "attend-handover-meeting")]
     public bool? AttendHandoverMeeting { get; set; }
 
+    [BindProperty]
+    public Guid? TasksDataId { get; set; }
+    [BindProperty]
+    public ProjectType? Type { get; set; }
+
     public override async Task<IActionResult> OnGetAsync()
     {
-        await base.OnGetAsync(); 
+        await base.OnGetAsync();
+        Type = Project.Type;
+        TasksDataId = Project.TasksDataId?.Value;
+        if (Project.Type == ProjectType.Transfer)
+        {
+            ReviewProjectInformation = TransferTaskData.HandoverReview;
+            MakeNotes = TransferTaskData.HandoverNotes;
+            AttendHandoverMeeting = TransferTaskData.HandoverMeeting;
+            NotApplicable = TransferTaskData.HandoverNotApplicable;
+        }
+        else
+        {
+            ReviewProjectInformation = ConversionTaskData.HandoverReview;
+            MakeNotes = ConversionTaskData.HandoverNotes;
+            AttendHandoverMeeting = ConversionTaskData.HandoverMeeting;
+            NotApplicable = ConversionTaskData.HandoverNotApplicable;
+        }
         return Page();
+    } 
+
+    public async Task<IActionResult> OnPost()
+    {
+        await sender.Send(new UpdateHandoverWithDeliveryOfficerCommand(new TaskDataId(TasksDataId.GetValueOrDefault())!, Type, NotApplicable, ReviewProjectInformation, MakeNotes, AttendHandoverMeeting));
+        TempData.SetNotification(NotificationType.Success, "Success", "Task updated successfully");
+        return Redirect(string.Format(RouteConstants.ProjectTaskList, ProjectId));
     }
 }